Создал util.js и data.js подключил точку входа в HTML и сделал импорт рандомайзера
